# packages/ahg-ric/resources/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RIC-O Linked Data API",
    "description": "Records in Contexts (RIC-O) ISO 23278 Linked Data API for Heratio archival management system",
    "version": "1.0.0",
    "contact": {
      "name": "Johan Pieterse",
      "email": "johan@plainsailingisystems.co.za"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0"
    }
  },
  "servers": [
    {
      "url": "/api/ric/v1",
      "description": "Heratio RIC-O API"
    }
  ],
  "paths": {
...
    "/sparql": {
      "get": {
        "summary": "Execute SPARQL 1.1 query (GET)",
        "description": "Read-only SPARQL Query passthrough to the backing triplestore. Implements the openric-spec `sparql-access` Draft profile v0.1.0. Update operations (INSERT/DELETE/CLEAR/LOAD/DROP/CREATE/COPY/MOVE/ADD) are rejected with 403 + application/problem+json. Rate-limited at 60/minute/IP. Backend query timeout 30s.",
        "tags": ["SPARQL", "sparql-access"],
        "parameters": [
          {"name": "query", "in": "query", "required": true, "schema": {"type": "string"}, "description": "SPARQL 1.1 query string (SELECT / ASK / CONSTRUCT / DESCRIBE)."}
        ],
        "responses": {
          "200": {
            "description": "SPARQL 1.1 results. Content-Type is propagated from the backend per the request Accept header — application/sparql-results+json (default for SELECT/ASK), application/sparql-results+xml, text/csv, text/tab-separated-values, text/turtle (CONSTRUCT/DESCRIBE), application/ld+json (CONSTRUCT/DESCRIBE).",
            "content": {
              "application/sparql-results+json": {"schema": {"type": "object", "properties": {"head": {"type": "object"}, "results": {"type": "object"}}}}
            }
          },
          "400": {"description": "Malformed query (backend parse error). application/problem+json."},
          "403": {"description": "SPARQL Update operation rejected. application/problem+json with type=https://openric.org/errors/update-not-permitted."},
          "429": {"description": "Rate limit exceeded (60/minute/IP)."},
          "502": {"description": "SPARQL backend unavailable. application/problem+json."}
        }
      },
      "post": {
        "summary": "Execute SPARQL 1.1 query (POST)",
        "description": "Same surface as GET. Accepts the query either as `?query=` form field (Content-Type: application/x-www-form-urlencoded) or as the raw request body (Content-Type: application/sparql-query).",
        "tags": ["SPARQL", "sparql-access"],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {"schema": {"type": "object", "properties": {"query": {"type": "string"}}, "required": ["query"]}},
            "application/sparql-query": {"schema": {"type": "string"}}
          }
        },
        "responses": {
          "200": {"description": "Same as GET /sparql."},
          "400": {"description": "Malformed query."},
          "403": {"description": "SPARQL Update rejected."},
          "429": {"description": "Rate limit exceeded."},
          "502": {"description": "Backend unavailable."}
        }
      }
    },
...
  },
...
}